chore(evals,ci): relocate HITL SDK tests and lock eval CI
